# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "My Store",
      "description": "Juan Perez is a very popular guy here in my country. So I decided to create an account: id=jperez password=123456. The account has the 'guest' role which means than. Once you login with that id and password I recommend don't refresh the page. If you do, you'll have to login one more time",
      "contact": {
        "name": "Sandro Simón",
        "email": "hello@sandrosimon.com"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api/v1",
        "description": "Server"
      }
      
    ],
...
    "paths": {
      "/auth/login": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Get your access token here",
          "description": "Get your access token here",
          "operationId": "handleLogin",
          "requestBody": {
            "description": "Get your access token here",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "jperez"
                    },
                    "password": {
                      "type": "string",
                      "example": "123456"
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "jperez"
                      },
                      "password": {
                        "type": "string",
                        "example": "123456"
                      }
                    } 
                  }
                }
                
              }
            },
            "405": {
              "description": "Invalid input"
            }
          }
        }      
      },
      "/auth/changePassword": {
        "put": {
          "tags": [
            "Authentication"
          ],
          "summary": "Get your access token here",
          "description": "Get your access token here",
          "operationId": "handleChangePassword",
          "requestBody": {
            "description": "Get your access token here",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "currentPassword": {
                      "type": "string",
                      "example": "123456"
                    },
                    "newPassword": {
                      "type": "string",
                      "example": "123456"
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "jperez"
                      },
                      "password": {
                        "type": "string",
                        "example": "123456"
                      }
                    } 
                  }
                }
                
              }
            },
            "405": {
              "description": "Invalid input"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }      
      },
...
      "/customers": {
        "get": {
          "tags": ["Customer"],
          "summary": "User roles with access to this endpoint: admin, seller, guest",
          "description": "Get all travels",
          "operationId": "getAllCustomers",
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref":"#/components/schemas/Customer"
                    }  
                  }
                }
              }
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "post": {
          "tags": [
            "Customer"
          ],
          "summary": "User roles with access to this endpoint: admin, seller",
          "description": "Add a new travel to the database",
          "operationId": "createOneCustomer",
          "requestBody": {
            "description": "Create a new travel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Customer"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/Customer"  
                  }
                }
                
              }
            },
            "405": {
              "description": "Invalid input"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }      
      },
      "/customers/{id}": {
        "get": {
          "tags": ["Customer"],
          "summary": "User roles with access to this endpoint: admin, seller, guest",
          "description": "Returns a travel",
          "operationId": "getOneCustomer",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Code of the account to return",
              "required": true,
              "schema": {
                "type": "string",
                "pattern": "^[0-9]{8,10}$",
                "example": "12345678"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/Customer"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Account not found"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "put": {
          "tags": ["Customer"],
          "summary": "User roles with access to this endpoint: admin, seller",
          "description": "Update an existing travel by Id",
          "operationId": "updateCustomer",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Code of the travel to return",
              "required": true,
              "schema": {
                "type": "string",
                "pattern": "^[0-9]{8,10}$",
                "example": "12345678"
              }
            } 
          ],
          "requestBody": {
            "description": "Update an existent travel in the database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                },
                "example": {
                  "email": "mfragides@gmail.com",
	                "firstName": "MISHEL",
                  "lastName": "FRAGIDES",
                  "salespersonId": "ssimon"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid account code supplied"
            },
            "404": {
              "description": "Account not found"
            },
            "405": {
              "description": "Validation exception"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
          
        },
        "delete": {
          "tags": ["Customer"],
          "summary": "User roles with access to this endpoint: admin",
          "description": "delete an account",
          "operationId": "deleteCustomer",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The customer id to remove",
              "required": true,
              "schema": {
                "type": "string",
                "pattern": "^[0-9]{8,10}$",
                "example": "12345678"
              }
            }
          ],
          "responses": {
            "204":{
              "description": "The account was deleted successfully."
            },
            "500":{
              "description": "Unexpected error."
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Customer": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "40483785"
            },
            "email": {
              "type": "string",
              "example": "lgalindo@mail.com"
            },
            "firstName": {
              "type": "string",
              "example": "Marylin"
            },
            "lastName": {
              "type": "string",
              "example": "Galindo"
            },
            "salespersonId": {
              "type": "string",
              "example": "mgalindo"
            }
          }
        },
...
      },
      "securitySchemes": {
        "storeAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }  
  }